pyproject.toml
[tool.poetry]
name = "pet-project"
version = "0.1.0"
description = ""
authors = ["S1m-s1m <139076421+S1m-s1m@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest-django = "^4.8.0"
celery = "^5.4.0"
psycopg2 = "^2.9.9"
django-parler = "^2.3"
rosetta = "^0.3"
stripe = "^10.4.0"
weasyprint = "^62.3"
pytest = "^8.3.1"
transliterate = "^1.10.2"
python-slugify = "^8.0.4"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
scipy = "^1.14.0"
nltk = "^3.8.1"
django = "^5.0.7"
pandas = "^2.2.2"
redis = "^5.0.7"
cffi = "^1.16.0"
pillow = "^10.4.0"
django-rosetta = "^0.10.0"
eventlet = "^0.36.1"
factory-boy = "^3.3.0"
mixer = "^7.2.2"


[tool.poetry.group.dev.dependencies]
pylint-django = "^2.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
